Refactor auth visitor to accept static_variant directly #1674
