+++ libraries/protocol/include/steem/protocol/operation_util_impl.hpp
@@ -46,32 +46,6 @@ struct operation_validate_visitor
    void operator()( const T& v )const { v.validate(); }
 };
 
-struct operation_get_required_auth_visitor
-{
-   typedef void result_type;
-
-   flat_set< account_name_type >&        active;
-   flat_set< account_name_type >&        owner;
-   flat_set< account_name_type >&        posting;
-   std::vector< authority >&  other;
-
-   operation_get_required_auth_visitor(
-         flat_set< account_name_type >& a,
-         flat_set< account_name_type >& own,
-         flat_set< account_name_type >& post,
-         std::vector< authority >& oth )
-      : active( a ), owner( own ), posting( post ), other( oth ) {}
-
-   template< typename T >
-   void operator()( const T& v )const
-   {
-      v.get_required_active_authorities( active );
-      v.get_required_owner_authorities( owner );
-      v.get_required_posting_authorities( posting );
-      v.get_required_authorities( other );
-   }
-};
-
 } } // steem::protocol
 
 //
@@ -129,7 +103,7 @@ void operation_get_required_authorities( const OperationType& op,          \
                                          flat_set< account_name_type >& posting,        \
                                          std::vector< authority >& other )     \
 {                                                                          \
-   op.visit( steem::protocol::operation_get_required_auth_visitor( active, owner, posting, other ) ); \
+   op.visit( steem::protocol::get_required_auth_visitor( active, owner, posting, other ) ); \
 }                                                                          \
                                                                            \
 } } /* steem::protocol */
